Table.cs extraction from TableReservation.cs Added field vietos to Table object
