+++ CafeApp/CafeApp/TableReservation.cs
@@ -10,46 +10,6 @@ using System.Windows.Forms;
 
 namespace CafeApp
 {
-    public class Table
-    {
-        public int _id;
-        public string _kavinesPavadinimas;
-
-        //ISO 8601 format
-        //YYYY-MM-DDThh:mm:ss
-        //The T indicates the start of the time part of the datetime value.
-        public DateTime _reservedOnFullDate;   
-        public TimeSpan _reservedOnHours;
-
-        public Table(int id, string kavinespavadinimas, DateTime reservedOnFullDate) //Pretty much just a testing constructor
-        {
-            _id = id;
-            _kavinesPavadinimas = kavinespavadinimas;
-            ChangeReservedDate(reservedOnFullDate);
-        }
-
-        public Table(int id, string kavinesPavadinimas) //The regular constructor
-        {
-            _id = id;
-            _kavinesPavadinimas = kavinesPavadinimas;
-        }
-
-        public void ChangeReservedDate(DateTime newDate)
-        {
-            _reservedOnFullDate = newDate;
-            _reservedOnHours = _reservedOnFullDate.TimeOfDay;
-        }
-
-        void ChangeReservedTime(TimeSpan newTime)
-        {
-            _reservedOnHours = newTime;
-        }
-
-        void addTableToDatabase()
-        {
-
-        }
-    }
 
     class TableReservation
     {
@@ -66,7 +26,6 @@ namespace CafeApp
             GetAllTables(kavinesPavadinimas);
         }
 
-
         public void GetAllTables(string kavinesPavadinimas)
         {
             //connection.Open(); //Bug: connection is not closed
@@ -82,7 +41,7 @@ namespace CafeApp
                 {
                     int readId = sqlDataReader.GetInt32(0); //Id column
                     string readKavine = sqlDataReader.GetString(1);
-                    DateTime readRezervuota = DateTime.MinValue; //Can't be NULL, so we'll assume min value is NULL
+                    DateTime readRezervuota = DateTime.MinValue; //Can't be assigned NULL, so we'll assume min value is NULL
                     if (!sqlDataReader.IsDBNull(2))
                         readRezervuota = sqlDataReader.GetDateTime(2);
